requirements/base.txt
Django==3.2.12
django-cors-headers==3.11.0
django-extensions==3.1.5
django-q==1.3.9
djangorestframework==3.13.1
# djangorestframework-jwt==1.11.0 # Remove this dependency
djangorestframework-simplejwt==5.0.0
django-filter==21.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2021.3
Pillow==9.0.1
openpyxl==3.0.9
redis==3.5.3
django-redis==5.2.0
Markdown==3.3.6
xhtml2pdf==0.2.5
reportlab==3.6.6   # issue due to xhtml2pdf inter-dependency reportlab, works perfectly on this version
git+https://github.com/kurenai-ryu/pyzk.git
sorl-thumbnail==12.8.0
django-cuser==2017.3.16
idna==2.10
Twisted[tls,http2]==22.4.0
autobahn==21.2.1
pyOpenSSL==19.1.0
channels==3.0.4
channels-redis==3.3.1
daphne==3.0.2
django-import-export==2.7.1
tablib==3.2.1
bs4==0.0.1
svgwrite==1.3.1
svgutils==0.3.1
cryptography==2.9.2
pyjwkest==1.4.2
django-oauth-toolkit==1.2.0
lxml==4.4.1
django-admin-rangefilter==0.8.4
requests==2.25.1
python-dotenv==0.17.0
PGPy==0.5.4
hkp4py==0.2.3.1
git+https://github.com/aayulogic/payroll-plugin-builder.git
regex==2022.1.18
python-magic==0.4.24
jsonschema==4.1.1
wheel
# for ADMS connection / Compilation
Cython==0.29.16
# mysqlclient==1.4.6
# PyMySQL==0.9.3
python-docx==0.8.11
html2text==2020.1.16
fast-map==0.1.0
